final_code/requirements.txt
streamlit==0.80.0
web3==5.17.0
Pillow==8.2.0
python-dotenv==0.17.1
